requirements.txt
# SafeTravels API - Requirements (RAG-focused)

# API Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# RAG Pipeline
langchain>=0.1.0
langchain-openai>=0.0.5
langgraph>=0.0.10
fastmcp>=2.0
chromadb>=0.4.0
sentence-transformers>=2.2.0

# LLM
openai>=1.0.0

# Database
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0

# Geospatial
geopy>=2.4.0

# HTTP Client
httpx>=0.25.0
requests>=2.31.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Utilities
python-dotenv>=1.0.0

# Web Scraping (for data ingestion)
beautifulsoup4>=4.12.0

# Dashboard
streamlit>=1.28.0
folium>=0.15.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
